feat(linter): Common logic extract to common trait, and common diagnostic messages and methods
