+++ crates/oxc_linter/src/utils/jest_vitest_rules/mod.rs
@@ -1,0 +1,1 @@
+pub mod prefer_expect_assertions;
